feature: добавил jdbc репозитории
так же обновил модели для реализации soft-delete в базе данных

+++ scrapper/src/main/java/backend/academy/scrapper/repository/link/LinkRepository.java
@@ -17,7 +17,7 @@ public interface LinkRepository {
 
     CompletableFuture<Void> update(Link link);
 
-    CompletableFuture<Void> deleteById(int id);
+    CompletableFuture<Void> deleteById(long id);
 
     CompletableFuture<Void> delete(Link link);
 }
